narrowing conversation fixed

+++ tasks/image_processor/negative.cpp
@@ -4,9 +4,9 @@ Image Negative::Apply(const Image& image) {
     Image result(image.Height(), image.Width());
     for (size_t i = 0; i < result.Height(); ++i) {
         for (size_t j = 0; j < result.Width(); ++j) {
-            result[i][j].Red = COLOR_MAX - image[i][j].Red;
-            result[i][j].Green = COLOR_MAX - image[i][j].Green;
-            result[i][j].Blue = COLOR_MAX - image[i][j].Blue;
+            result[i][j].Red = static_cast<int16_t>(COLOR_MAX - image[i][j].Red);
+            result[i][j].Green = static_cast<int16_t>(COLOR_MAX - image[i][j].Green);
+            result[i][j].Blue = static_cast<int16_t>(COLOR_MAX - image[i][j].Blue);
         }
     }
     return result;
